Update HIRES benchmark
